Downloading csv file from yahoo finance.
Class YahooLinkGen: Generates the URI to download the yahoo finance history.
FileDownloader: this class can download any file, but in this case help us download the csv from yahoo finance.
OutputType: a constant holder to help us decide the output format of the csv file and what data to get, daily or monthly.